Responding to Leyla's comments
