+++ cloud/aws/templates/aws_oidc/bin/setup.py
@@ -3,7 +3,7 @@ import subprocess
 import shlex
 import shutil
 
-from cloud.aws.templates.aws_oidc.bin.awsSetupTemplate import AwsSetupTemplate
+from cloud.aws.templates.aws_oidc.bin.aws_template import AwsSetupTemplate
 from cloud.aws.bin.lib import backend_setup
